% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feast_core.R
\name{.compute_loglik}
\alias{.compute_loglik}
\title{Compute mixture log-likelihood}
\usage{
.compute_loglik(y_sink, x_source, rho, p_unknown)
}
\arguments{
\item{y_sink}{Numeric vector.}

\item{x_source}{Matrix, n_otus x K.}

\item{rho}{Numeric vector, length K+1.}

\item{p_unknown}{Numeric vector, length n_otus.}
}
\value{
Scalar log-likelihood value.
}
\description{
Used as the convergence criterion in .run_EM() [5].
}
\keyword{internal}
