% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feast_core.R
\name{.E_step}
\alias{.E_step}
\title{E-step: compute posterior latent source assignments}
\usage{
.E_step(y_sink, x_source, rho, p_unknown)
}
\arguments{
\item{y_sink}{Numeric vector, length n_otus. Sink proportions.}

\item{x_source}{Matrix, n_otus x K. Source profiles.}

\item{rho}{Numeric vector, length K+1. Mixing proportions.}

\item{p_unknown}{Numeric vector, length n_otus. Unknown source profile.}
}
\value{
Matrix, n_otus x (K+1). Row-normalised posterior weights.
}
\description{
E-step: compute posterior latent source assignments
}
\keyword{internal}
