% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feast_core.R
\name{.init_params}
\alias{.init_params}
\title{Initialise EM parameters}
\usage{
.init_params(n_sources, n_otus, y_sink, x_source)
}
\arguments{
\item{n_sources}{Integer. Number of known sources.}

\item{n_otus}{Integer. Number of OTUs.}

\item{y_sink}{Numeric vector, length n_otus. Sink proportions.}

\item{x_source}{Matrix, n_otus x K. Source profiles.}
}
\value{
List: \code{rho} (length K+1), \code{p_unknown} (length n_otus).
}
\description{
Mixing proportions (rho) are random uniform.
Unknown source profile uses biologically-informed initialisation [2].
}
\keyword{internal}
