% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEASTime-package.R
\docType{package}
\name{FEASTime}
\alias{FEASTime-package}
\alias{FEASTime}
\title{FEASTime: Time-Series Microbial Source Tracking}
\description{
Extends FEAST to multi-timepoint longitudinal data via a
Markov chain-based iterative EM framework.
}
\section{Main function}{

\code{\link{run_feastime}}
}

\section{Visualisation}{

\code{\link{plot_contributions}}, \code{\link{plot_trends}}
}

\section{Utilities}{

\code{\link{summarise_feastime}}, \code{\link{export_xlsx}}
}

\references{
Shenhav et al. (2019). FEAST: fast expectation-maximization for
microbial source tracking. Nature Methods, 16, 627-632.
}
\author{
\strong{Maintainer}: Ma Yanyan \email{ma15091227920@163.com}

}
